# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber.uz CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: string;
  barberId: string;
  clientId?: string | null;
  clientName: string;
  serviceId?: string | null;
  serviceName?: string | null;
  date: Date;
  startTime: string;
  endTime: string;
  price: number;
  status: BookingStatus;
  notes?: string | null;
  createdAt: Date;
}