# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber.uz CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingRequest {
  clientId?: string | null;
  clientName: string;
  serviceId?: string | null;
  date: Date;
  startTime: string;
  endTime: string;
  price: number;
  notes?: string | null;
}