# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Barber.uz CRM API
 * OpenAPI spec version: 0.1.0
 */
import type { UserLang } from "./userLang";
import type { UserMode } from "./userMode";

export interface User {
  id: string;
  name: string;
  username: string;
  brandName?: string | null;
  mode: UserMode;
  lang: UserLang;
  telegramVerified: boolean;
  telegramId?: string | null;
  createdAt: Date;
}